+++ lib/controller/useful_widgets.dart
@@ -1,3 +1,4 @@
+import 'package:email_validator/email_validator.dart';
 import 'package:flutter/material.dart';
 import 'package:instagram_flutter/models/user.dart';
 
@@ -88,5 +89,74 @@ Future<void> delUserAlert(id, context) async {
         duration: Duration(seconds: 5),
         backgroundColor: Colors.white,
         content: Text('Sua conta foi excluída permanentemente.')));
+  }
+}
+
+class FormFieldWithPadding extends StatelessWidget {
+  final TextEditingController? controller;
+  final String? initialValue;
+  final String validatorReturn;
+  final String hintTextDecoration;
+  final TextInputType textInputType;
+  final TextCapitalization textCapitalization;
+  final bool isPassword;
+  final bool isEmail;
+  final bool isUsername;
+
+  const FormFieldWithPadding(
+      {Key? key,
+      this.controller,
+      required this.validatorReturn,
+      required this.hintTextDecoration,
+      required this.textInputType,
+      required this.textCapitalization,
+      required this.isPassword,
+      required this.isEmail,
+      required this.isUsername,
+      this.initialValue})
+      : super(key: key);
+
+  bool emailValidator() {
+    return EmailValidator.validate(controller!.text);
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Padding(
+      padding: const EdgeInsets.only(right: 16.0, left: 16.0, top: 10.0),
+      child: TextFormField(
+          controller: controller,
+          obscureText: isPassword,
+          keyboardType: textInputType,
+          textCapitalization: textCapitalization,
+          initialValue: initialValue,
+          decoration: InputDecoration(
+            border: const OutlineInputBorder(),
+            label: Text(hintTextDecoration),
+          ),
+          validator: isEmail
+              ? (value) {
+                  if (!emailValidator()) {
+                    return validatorReturn;
+                  }
+                  return null;
+                }
+              : isUsername
+                  ? (value) {
+                      if (value == null ||
+                          value.isEmpty ||
+                          value.contains(' ') ||
+                          value.length > 20) {
+                        return validatorReturn;
+                      }
+                      return null;
+                    }
+                  : (value) {
+                      if (value == null || value.isEmpty || value.length > 50) {
+                        return validatorReturn;
+                      }
+                      return null;
+                    }),
+    );
   }
 }
